donezo

+++ src/board/board.rs
@@ -206,11 +206,11 @@ impl Board {
     }
 
     fn material_val(&self, c: Color) -> i32 {
-        self.bitboard(c, PieceName::Queen).count_bits() * PieceName::Queen.value()
-            + self.bitboard(c, PieceName::Rook).count_bits() * PieceName::Rook.value()
-            + self.bitboard(c, PieceName::Bishop).count_bits() * PieceName::Bishop.value()
-            + self.bitboard(c, PieceName::Knight).count_bits() * PieceName::Knight.value()
-            + self.bitboard(c, PieceName::Pawn).count_bits() * PieceName::Pawn.value()
+        self.bitboard(c, PieceName::Queen).count_bits() as i32 * PieceName::Queen.value()
+            + self.bitboard(c, PieceName::Rook).count_bits() as i32 * PieceName::Rook.value()
+            + self.bitboard(c, PieceName::Bishop).count_bits() as i32 * PieceName::Bishop.value()
+            + self.bitboard(c, PieceName::Knight).count_bits() as i32 * PieceName::Knight.value()
+            + self.bitboard(c, PieceName::Pawn).count_bits() as i32 * PieceName::Pawn.value()
     }
 
     pub fn material_balance(&self) -> i32 {
